Move Presto Spark execution exceptions to presto-spark-classloader-interface
As they have to be serializable
