+++ presto-spark-base/src/test/java/com/facebook/presto/spark/execution/TestPrestoSparkExecutionExceptionFactory.java
@@ -15,6 +15,7 @@ package com.facebook.presto.spark.execution;
 
 import com.facebook.presto.execution.ExecutionFailureInfo;
 import com.facebook.presto.execution.Failure;
+import com.facebook.presto.spark.classloader_interface.PrestoSparkExecutionException;
 import com.facebook.presto.spi.ErrorCode;
 import com.facebook.presto.spi.PrestoException;
 import com.google.common.collect.ImmutableList;
